Casting issue fixed

+++ src/libopensc/reader-pcsc.c
@@ -341,7 +341,7 @@ static int refresh_attributes(sc_reader_t *reader)
 		}
 		
 		/* the system could not dectect any reader. It means, the prevoiusly attached reader is disconnected. */		
-		if (rv == SCARD_E_NO_READERS_AVAILABLE || rv == SCARD_E_SERVICE_STOPPED) {
+		if (rv == (LONG)SCARD_E_NO_READERS_AVAILABLE || rv == (LONG)SCARD_E_SERVICE_STOPPED) {
  			if (old_flags & SC_READER_CARD_PRESENT) {
  				reader->flags |= SC_READER_CARD_CHANGED;
  			}
